Implements pipelining file I/O (similar to tail -f or unix pipes)
For ID files being downloaded from another host, block on sfs_getattr
and sfs_read until enough of the file has been downloaded to satisfy the
next read.  This ensures the reader doesn't get an EOF, and will keep
reading rather than closing the file.

Need to block on sfs_getattr because it seems like some processes (e.g.,
tail) will use fstat to check the size of the file and then won't read
beyond that, even if between the stat and read more data gets
downloaded.

We really need stat to report a size past the end of the process' next
read.  E.g., if the next read will be 4 bytes starting at byte 8, the
size needs to be > 12 so the process will need to issue another read.
Otherwise, the process assumes the last read reached the EOF.

Since we don't know how large the next read will be when sfs_getattr is
called, for now just assume the read will be <= BLKSIZE.  So we should
wait to return from sfs_getattr until size is > (current pos + BLKSIZE).
I don't know if that's possible, since 'current pos' is a per-process
thing right?