Use one f3 driver instance for every accessor, track open files with textfiles
We were blocking on returning from sfs_read until more data was written or writer
exited.  If reader and writer are on the same host though the writer can't
progress (since blocking in sfs_read ties up the fuse thread needed for the
writer).  Instead use a separate instance of f3 for each accessor, so
the reader waiting doesn't block the writer.

When the writer is remote, the reader can check if the download
connection is still active to see if the writer is still writing.
When reader and writer are on the same host, need to use another
mechanism - for now, just use text files to track files that are
open for writing and files that are done being written to.